# internal/http/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "rssam API",
    "version": "1.0.0",
    "description": "REST API of the rssam RSS aggregator.\n\n**Authentication.** Every `/v1` endpoint takes `X-Auth-Token: <token>` where the token is either a personal API key (`POST /v1/me/api-keys`) or the global `AUTH_TOKEN` (maps to user id 1). A browser session cookie is also accepted, but state-changing requests authenticated only by the cookie must pass the same-origin check (`Sec-Fetch-Site`/`Origin`), otherwise 403.\n\n**Envelope.** Successful JSON responses are `{\"data\": ..., \"total\": n}`; errors are `{\"error_message\": \"...\"}`. Lists take `limit`/`offset`; `total` is the full row count.\n\n**Ownership.** Objects are scoped to the calling user; reading another user's object yields 404. Endpoints marked \"Requires an admin principal\" return 403 for non-admins.\n\n**Request id.** Every response carries `X-Request-Id` (echoed from the request when a proxy set it, generated otherwise); quote it when reporting 5xx errors.\n\n**Limits.** Request bodies are capped (MAX_REQUEST_BODY_BYTES, default 1 MiB) → 413; per-IP rate limits on authentication failures and sensitive routes → 429."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "security": [
    {
      "AuthToken": []
    },
    {
      "SessionCookie": []
    }
  ],
  "paths": {
...
    "/v1/feeds/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "description": "Object id."
        }
      ],
      "get": {
        "tags": [
          "feeds"
        ],
        "summary": "Get feed",
        "operationId": "getFeed",
        "responses": {
          "200": {
            "description": "Feed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "put": {
        "tags": [
          "feeds"
        ],
        "summary": "Update feed",
        "operationId": "updateFeed",
        "description": "Requires an admin principal.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Another feed already uses this URL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "delete": {
        "tags": [
          "feeds"
        ],
        "summary": "Delete feed",
        "operationId": "deleteFeed",
        "description": "Requires an admin principal.",
        "responses": {
          "200": {
            "description": "Deleted with its entries.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auth-Token",
        "description": "Personal API key or the global AUTH_TOKEN."
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id",
        "description": "Web UI session. Mutations additionally require a same-origin request."
      },
      "MetricsBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "METRICS_TOKEN for /metrics only."
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Malformed input or validation error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credentials.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Admin required, or a cookie-authenticated request that failed the same-origin check.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Object not found or not owned by the caller.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected server error. The response carries an X-Request-Id header to correlate with server logs.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "The backing store for this resource is not configured.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error_message": {
            "type": "string"
          }
        },
        "required": [
          "error_message"
        ],
        "description": "Error envelope returned by every non-2xx JSON response."
      },
      "Deleted": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ]
      },
...
      "Feed": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "feed_url": {
            "type": "string",
            "format": "uri"
          },
          "feed_type": {
            "type": "string",
            "description": "Detected feed format (rss, atom, json, ...); omitted when unknown."
          },
          "title": {
            "type": "string"
          },
          "category_id": {
            "type": "integer",
            "format": "int64"
          },
          "interval_minutes": {
            "type": "integer",
            "description": "Refresh interval; 0 = use the global default."
          },
          "scraper_rules": {
            "type": "string"
          },
          "rewrite_rules": {
            "type": "string"
          },
          "blocked_rules": {
            "type": "string"
          },
          "keep_rules": {
            "type": "string"
          },
          "fetch_via_proxy": {
            "type": "boolean"
          },
          "tls_insecure": {
            "type": "boolean"
          },
          "crawler": {
            "type": "boolean"
          },
          "user_agent": {
            "type": "string"
          },
          "store_hash_only": {
            "type": "boolean"
          },
          "entry_retention_days": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "feed_url",
          "title",
          "interval_minutes",
          "fetch_via_proxy",
          "tls_insecure",
          "crawler"
        ]
      },
      "FeedWrite": {
        "type": "object",
        "properties": {
          "feed_url": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "category_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "interval_minutes": {
            "type": "integer",
            "description": "Refresh interval; 0 = use the global default."
          },
          "scraper_rules": {
            "type": "string"
          },
          "rewrite_rules": {
            "type": "string"
          },
          "blocked_rules": {
            "type": "string"
          },
          "keep_rules": {
            "type": "string"
          },
          "fetch_via_proxy": {
            "type": "boolean"
          },
          "tls_insecure": {
            "type": "boolean"
          },
          "crawler": {
            "type": "boolean"
          },
          "user_agent": {
            "type": "string"
          },
          "store_hash_only": {
            "type": "boolean"
          },
          "entry_retention_days": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": [
          "feed_url"
        ],
        "description": "Full representation on PUT: omitted fields are reset to their zero value."
      },
...
      "FeedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Feed"
          },
          "total": {
            "type": "integer",
            "description": "Number of items in data (1 for single objects) or total rows for lists."
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
...
      "DeletedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Deleted"
          },
          "total": {
            "type": "integer",
            "description": "Number of items in data (1 for single objects) or total rows for lists."
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
...
    }
  }
}